+++ example/lib/cards/cards_screen.dart
@@ -27,8 +27,8 @@ class _CardsScreenState extends State<CardsScreen>
   @override
   void initState() {
     super.initState();
-    cards.setAppOpenCardsSyncListener((moe.SyncCompleteData? data) {
-      debugPrint('Cards App Open Sync Listener Callback: $data');
+    cards.setSyncCompleteListener((moe.SyncCompleteData? data) {
+      debugPrint('Cards Sync Listener Callback: $data');
     });
     setUpTabs();
     cards.onCardsSectionLoaded((moe.SyncCompleteData? data) {