+++ core/moengage_flutter_android/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Next Version
 - Bundling `moe-android-sdk` & `inapp` dependencies with the plugin 
+- `moe-android-sdk` version updated to `13.05.00`
+- `inapp` version updated to `8.6.0`
 
 # 07-08-2024